Prototype of user defined artist role for display in Chromecast
LMS v9 introduces user defined roles, and user defined roles are
returned via the "all roles" tag (`A` rather than `a`) of the CLI, so as
a quick test try to replace standard artist with my "DisplayArtist" tag.
This works for my library, but hasn't been tested significantly.  Long
term the actual role used should be an option in the plugin settings.